+++ CHANGELOG.md
@@ -7,6 +7,19 @@ Most recent releases are shown at the top. Each release shows:
 - **Fixed**: Bug fixes that don't change documented behaviour
 
 
+## 0.3.1 (2026-01-28)
+
+### new:
+- N/A
+
+### changed
+- N/A
+
+### fixed:
+- Fix `edit_file` but with trailing new lines (#28)
+
+
+
 ## 0.3.0 (2026-01-28)
 
 ### new: